refactor(backend): group runtime entrypoints by platform
- move Bun launchers under src/platforms/bun
- move Cloudflare worker, mounts, env, and backend composition under src/platforms/cloudflare
- update Alchemy, package scripts, tests, and lint overrides for the new paths
