+++ apps/backend/.oxlintrc.json
@@ -56,7 +56,7 @@
       }
     },
     {
-      "files": ["src/cloudflare/**/*.ts", "src/composition/**/*.ts"],
+      "files": ["src/platforms/cloudflare/**/*.ts"],
       "rules": {
         "no-restricted-imports": [
           "error",
@@ -76,7 +76,7 @@
       }
     },
     {
-      "files": ["src/bun/**/*.ts"],
+      "files": ["src/platforms/bun/**/*.ts"],
       "rules": {
         "no-restricted-imports": [
           "error",
@@ -93,7 +93,7 @@
       }
     },
     {
-      "files": ["src/bun/**/main.ts", "src/bun/**/*-main.ts"],
+      "files": ["src/platforms/bun/**/main.ts", "src/platforms/bun/**/*-main.ts"],
       "rules": {
         "no-restricted-imports": [
           "error",
