+++ apps/backend/package.json
@@ -8,11 +8,11 @@
   },
   "scripts": {
     "dev": "bun run http",
-    "http": "bun run src/bun/http-main.ts",
-    "cli": "bun run src/bun/cli-main.ts",
-    "mcp:stdio": "bun run src/bun/mcp-stdio-main.ts",
-    "mcp:http": "bun run src/bun/mcp-http-main.ts",
-    "dev:local:api": "bun run src/bun/local-backend-main.ts",
+    "http": "bun run src/platforms/bun/http-main.ts",
+    "cli": "bun run src/platforms/bun/cli-main.ts",
+    "mcp:stdio": "bun run src/platforms/bun/mcp-stdio-main.ts",
+    "mcp:http": "bun run src/platforms/bun/mcp-http-main.ts",
+    "dev:local:api": "bun run src/platforms/bun/local-backend-main.ts",
     "db:auth:schema": "bun run --cwd ../../packages/coffee/external/sqlite auth:schema",
     "db:check": "bun run --cwd ../../packages/coffee/external/sqlite db:check",
     "db:draft": "bun run --cwd ../../packages/coffee/external/sqlite db:draft",
