+++ apps/backend/src/platforms/bun/cli-main.ts
