+++ apps/backend/src/platforms/bun/mcp-stdio-main.ts
