+++ apps/backend/src/platforms/cloudflare/direct-http-auth.ts
