+++ apps/backend/src/platforms/cloudflare/mounts/mcp.ts
@@ -4,10 +4,7 @@ import {
   cloudflareResponse,
   type CloudflareMount,
 } from "@effect-coffee-shop/backend-host/mount";
-import {
-  createCloudflareRequestServices,
-  getCloudflareRuntimeBackend,
-} from "../../composition/coffee-backend.ts";
+import { createCloudflareRequestServices, getCloudflareRuntimeBackend } from "../coffee-backend.ts";
 import { systemActor } from "@effect-coffee-shop/coffee-core/application/CurrentActor";
 
 const isMcpRequest = (request: Request): boolean => cloudflarePathIsOrStartsWith(request, "/mcp");
